Publish artifacts in local Maven repository
Thus the artifacts can be used from clojure Leiningen projects, since
Leiningen uses maven for dependency grabing.

Because Maven doesn't support version ranges in local repositories,
the precise versions have to be specified in the ivy.xml files. An
alternative would be to use a branch of Cake that has ivy support.