Fix bug
The repeat? method was being applied to the pipe and branch
transformers but not to the standard transformers. Instead it was
applied to the container transformer. Now all methods that add a
transformer return the parent transformer. When executing the repeat?
transformer the parent applies the repeat clause to the last defined
transformer.